+++ python/cugraph/link_prediction/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2018-2019, NVIDIA CORPORATION.
+# Copyright (c) 2019, NVIDIA CORPORATION.
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
 # You may obtain a copy of the License at
@@ -11,17 +11,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# See the docstring in versioneer.py for instructions. Note that you must
-# re-run 'versioneer.py setup' after changing this section, and commit the
-# resulting files.
-
-[versioneer]
-VCS = git
-style = pep440
-versionfile_source = cugraph/_version.py
-versionfile_build = cugraph/_version.py
-tag_prefix = v
-parentdir_prefix = cugraph-
-
-[flake8]
-exclude = img,notebooks,thirdparty,__init__.py,libgdf
+from cugraph.link_prediction.jaccard import jaccard
+from cugraph.link_prediction.overlap import overlap
+from cugraph.link_prediction.wjaccard import jaccard_w
+from cugraph.link_prediction.woverlap import overlap_w