+++ CHANGELOG.md
@@ -3,10 +3,12 @@
 ## New Features
 - PR #361 Prototypes for cusort functions
 - PR #357 Pagerank cpp API
+- PR #366 Adds graph.degrees() function returning both in and out degree.
 
 
 ## Improvements
 - PR #353 Change snmg python wrapper in accordance to cpp api
+- PR #362 Restructured python/cython directories and files.
 - PR #365 Updates for setting device and vertex ids for snmg pagerank
 
 ## Bug Fixes